swap pycrypto with pycryptodome
pycrypto is a dead repo and hasn't been updated for more than 5 years.  pycryptodome is a fork of pycrypto that's still actively maintained.  It's namespace structure matches that of the original pycrypto, so it's considered backward compatible.  The driver for this change is that pycrypto fails during build on newer Windows/Visual Studio installations due to out of date requirements.